chore: create handle_parent_operator function
